README: add cached credentials howto
Add a subsection to the configuration section for a cached credentials
setup howto using pam_ccreds. This can be used for implementing LDAP
"offline authentification" using pam_ldapdb.

As it took me some time to find a working configuration it may be
helpful for others which are facing the same requirements.

Signed-off-by: Richard Leitner <richard.leitner@skidata.com>